jQuery hides flash notice of order confirmation after an item is added to cart
When a user completes the checkout process they are presented with a
flash notice that confirms the checkout.  If they have javascript
enabled, and they add an item to the cart, the cart refreshes via an
ajax call.  This typically leaves the flash notice within the main body
of the page - but now we hide it using jQuery once an item is added to
the cart.